delete-app-revision: fix review
