+++ pkg/resourcekeeper/gc.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"encoding/json"
 	"math/rand"
-	"sort"
 	"strings"
 	"time"
 
@@ -32,16 +31,13 @@ import (
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	ktypes "k8s.io/apimachinery/pkg/types"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
-	"k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/common"
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/v1alpha1"
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/v1beta1"
 	"github.com/oam-dev/kubevela/pkg/auth"
-	"github.com/oam-dev/kubevela/pkg/cache"
 	"github.com/oam-dev/kubevela/pkg/features"
 	"github.com/oam-dev/kubevela/pkg/monitor/metrics"
 	"github.com/oam-dev/kubevela/pkg/multicluster"
@@ -174,8 +170,8 @@ func (h *resourceKeeper) garbageCollect(ctx context.Context, cfg *gcConfig) (fin
 	}
 
 	if !cfg.disableApplicationRevisionGC {
-		if err = gc.GarbageCollectApplicationRevisionResourceTrackers(ctx); err != nil {
-			return false, waiting, errors.Wrapf(err, "failed to garbage collect application revision resource trackers")
+		if err = gc.GarbageCollectApplicationRevision(ctx); err != nil {
+			return false, waiting, errors.Wrapf(err, "failed to garbage collect application revision")
 		}
 	}
 
@@ -574,198 +570,4 @@ func (h *gcHandler) GarbageCollectLegacyResourceTrackers(ctx context.Context) er
 	}
 	h.app.ObjectMeta = app.ObjectMeta
 	return nil
-}
-
-// execute garbage collection functions, including:
-// - clean up legacy app revisions
-// - clean up legacy component revisions
-func (h *gcHandler) GarbageCollectApplicationRevisionResourceTrackers(ctx context.Context) error {
-	t := time.Now()
-	defer func() {
-		metrics.AppReconcileStageDurationHistogram.WithLabelValues("gc-app-rev").Observe(time.Since(t).Seconds())
-	}()
-	collectFuncs := []garbageCollectFunc{
-		garbageCollectFunc(cleanUpApplicationRevision),
-		garbageCollectFunc(cleanUpWorkflowComponentRevision),
-	}
-	for _, collectFunc := range collectFuncs {
-		if err := collectFunc(ctx, h); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-type garbageCollectFunc func(ctx context.Context, h *gcHandler) error
-
-// cleanUpApplicationRevision check all appRevisions of the application, remove them if the number of them exceed the limit
-func cleanUpApplicationRevision(ctx context.Context, h *gcHandler) error {
-	if h.cfg.disableApplicationRevisionGC {
-		return nil
-	}
-	t := time.Now()
-	defer func() {
-		metrics.AppReconcileStageDurationHistogram.WithLabelValues("gc-rev.apprev").Observe(time.Since(t).Seconds())
-	}()
-	sortedRevision, err := getSortedAppRevisions(ctx, h.Client, h.app.Name, h.app.Namespace)
-	if err != nil {
-		return err
-	}
-	appRevisionInUse := gatherUsingAppRevision(h.app)
-	appRevisionLimit := getApplicationRevisionLimitForApp(h.app, h.cfg.appRevisionLimit)
-	needKill := len(sortedRevision) - appRevisionLimit - len(appRevisionInUse)
-	if needKill <= 0 {
-		return nil
-	}
-	klog.InfoS("Going to garbage collect app revisions", "limit", h.cfg.appRevisionLimit,
-		"total", len(sortedRevision), "using", len(appRevisionInUse), "kill", needKill)
-
-	for _, rev := range sortedRevision {
-		if needKill <= 0 {
-			break
-		}
-		// don't delete app revision in use
-		if appRevisionInUse[rev.Name] {
-			continue
-		}
-		if err := h.Client.Delete(ctx, rev.DeepCopy()); err != nil && !kerrors.IsNotFound(err) {
-			return err
-		}
-		needKill--
-	}
-	return nil
-}
-
-func cleanUpWorkflowComponentRevision(ctx context.Context, h *gcHandler) error {
-	if h.cfg.disableComponentRevisionGC {
-		return nil
-	}
-	t := time.Now()
-	defer func() {
-		metrics.AppReconcileStageDurationHistogram.WithLabelValues("gc-rev.comprev").Observe(time.Since(t).Seconds())
-	}()
-	// collect component revision in use
-	compRevisionInUse := map[string]map[string]struct{}{}
-	ctx = auth.ContextWithUserInfo(ctx, h.app)
-	for i, resource := range h.app.Status.AppliedResources {
-		compName := resource.Name
-		ns := resource.Namespace
-		r := &unstructured.Unstructured{}
-		r.GetObjectKind().SetGroupVersionKind(resource.GroupVersionKind())
-		_ctx := multicluster.ContextWithClusterName(ctx, resource.Cluster)
-		err := h.Client.Get(_ctx, ktypes.NamespacedName{Name: compName, Namespace: ns}, r)
-		notFound := kerrors.IsNotFound(err)
-		if err != nil && !notFound {
-			return errors.WithMessagef(err, "get applied resource index=%d", i)
-		}
-		if compRevisionInUse[compName] == nil {
-			compRevisionInUse[compName] = map[string]struct{}{}
-		}
-		if notFound {
-			continue
-		}
-		compRevision, ok := r.GetLabels()[oam.LabelAppComponentRevision]
-		if ok {
-			compRevisionInUse[compName][compRevision] = struct{}{}
-		}
-	}
-
-	for _, curComp := range h.app.Status.AppliedResources {
-		crList := &appsv1.ControllerRevisionList{}
-		listOpts := []client.ListOption{client.MatchingLabels{
-			oam.LabelControllerRevisionComponent: utils.EscapeResourceNameToLabelValue(curComp.Name),
-		}, client.InNamespace(h.getComponentRevisionNamespace(ctx))}
-		_ctx := multicluster.ContextWithClusterName(ctx, curComp.Cluster)
-		if err := h.Client.List(_ctx, crList, listOpts...); err != nil {
-			return err
-		}
-		needKill := len(crList.Items) - h.cfg.appRevisionLimit - len(compRevisionInUse[curComp.Name])
-		if needKill < 1 {
-			continue
-		}
-		sortedRevision := crList.Items
-		sort.Sort(historiesByComponentRevision(sortedRevision))
-		for _, rev := range sortedRevision {
-			if needKill <= 0 {
-				break
-			}
-			if _, inUse := compRevisionInUse[curComp.Name][rev.Name]; inUse {
-				continue
-			}
-			_rev := rev.DeepCopy()
-			oam.SetCluster(_rev, curComp.Cluster)
-			if err := h.resourceKeeper.DeleteComponentRevision(_ctx, _rev); err != nil {
-				return err
-			}
-			needKill--
-		}
-	}
-	return nil
-}
-
-// gatherUsingAppRevision get all using appRevisions include app's status pointing to
-func gatherUsingAppRevision(app *v1beta1.Application) map[string]bool {
-	usingRevision := map[string]bool{}
-	if app.Status.LatestRevision != nil && len(app.Status.LatestRevision.Name) != 0 {
-		usingRevision[app.Status.LatestRevision.Name] = true
-	}
-	return usingRevision
-}
-
-func getApplicationRevisionLimitForApp(app *v1beta1.Application, fallback int) int {
-	for _, p := range app.Spec.Policies {
-		if p.Type == v1alpha1.GarbageCollectPolicyType && p.Properties != nil && p.Properties.Raw != nil {
-			prop := &v1alpha1.GarbageCollectPolicySpec{}
-			if err := json.Unmarshal(p.Properties.Raw, prop); err == nil && prop.ApplicationRevisionLimit != nil && *prop.ApplicationRevisionLimit >= 0 {
-				return *prop.ApplicationRevisionLimit
-			}
-		}
-	}
-	return fallback
-}
-
-// getSortedAppRevisions get application revisions by revision number
-func getSortedAppRevisions(ctx context.Context, cli client.Client, appName string, appNs string) ([]v1beta1.ApplicationRevision, error) {
-	revs, err := getAppRevisions(ctx, cli, appName, appNs)
-	if err != nil {
-		return nil, err
-	}
-	sort.Slice(revs, func(i, j int) bool {
-		ir, _ := util.ExtractRevisionNum(revs[i].Name, "-")
-		ij, _ := util.ExtractRevisionNum(revs[j].Name, "-")
-		return ir < ij
-	})
-	return revs, nil
-}
-
-// getAppRevisions get application revisions by label
-func getAppRevisions(ctx context.Context, cli client.Client, appName string, appNs string) ([]v1beta1.ApplicationRevision, error) {
-	appRevisionList := new(v1beta1.ApplicationRevisionList)
-	var err error
-	if cache.OptimizeListOp {
-		err = cli.List(ctx, appRevisionList, client.MatchingFields{cache.AppIndex: appNs + "/" + appName})
-	} else {
-		err = cli.List(ctx, appRevisionList, client.InNamespace(appNs), client.MatchingLabels{oam.LabelAppName: appName})
-	}
-	if err != nil {
-		return nil, err
-	}
-	return appRevisionList.Items, nil
-}
-
-func (h *gcHandler) getComponentRevisionNamespace(ctx context.Context) string {
-	if ns, ok := ctx.Value(0).(string); ok && ns != "" {
-		return ns
-	}
-	return h.app.Namespace
-}
-
-type historiesByComponentRevision []appsv1.ControllerRevision
-
-func (h historiesByComponentRevision) Len() int      { return len(h) }
-func (h historiesByComponentRevision) Swap(i, j int) { h[i], h[j] = h[j], h[i] }
-func (h historiesByComponentRevision) Less(i, j int) bool {
-	ir, _ := util.ExtractRevisionNum(h[i].Name, "-")
-	ij, _ := util.ExtractRevisionNum(h[j].Name, "-")
-	return ir < ij
 }
